Clean up queued retry code after persistent queue merge
